* Moved all JMX metrics to one common MBean * Renamed keys returned on statistics related metrics * Minor code changes
