+++ core/src/main/scala/org/elasticmq/metrics/MBeanMetrics.scala
@@ -1,11 +1,12 @@
 package org.elasticmq.metrics
 
-import java.util.concurrent.{ExecutorService, Executors, TimeUnit}
+import java.util.concurrent.{Executors, TimeUnit}
 
 import akka.actor.ActorRef
 import akka.util.Timeout
 import javax.management.openmbean._
 import org.elasticmq.QueueStatistics
+import org.elasticmq.metrics.QueuesMetrics.{queueDataNames, queueStatisticsCompositeType, queueStatisticsTabularType}
 import org.elasticmq.util.NowProvider
 
 import scala.concurrent.{Await, ExecutionContext}
@@ -13,35 +14,56 @@ import scala.concurrent.{Await, ExecutionContext}
 trait QueuesMetricsMBean {
   def getQueueNames: Array[String]
 
-  def deleteQueue(queueName: String): Unit
-}
-
-class QueuesMetrics(queueManagerActor: ActorRef) extends QueuesMetricsMBean {
-
-  implicit val timeout: Timeout = Timeout(21, TimeUnit.SECONDS)
-
-  override def getQueueNames: Array[String] = {
-    Await.result(QueuesMetricsOps.getQueueNames(queueManagerActor), timeout.duration).toArray
-  }
-
-  override def deleteQueue(queueName: String): Unit =
-    Await.ready(QueuesMetricsOps.deleteQueue(queueName, queueManagerActor), timeout.duration)
-
-}
-
-
-trait QueueMetricsMBean {
   def getNumberOfMessagesInQueue(queueName: String): CompositeData
 
   def getNumberOfMessagesForAllQueues: TabularData
 }
 
-class QueueMetrics(queueManagerActor: ActorRef) extends QueueMetricsMBean {
+class QueuesMetrics(queueManagerActor: ActorRef) extends QueuesMetricsMBean {
+
+  implicit val timeout: Timeout = Timeout(21, TimeUnit.SECONDS)
+  implicit val ec: ExecutionContext = ExecutionContext.fromExecutorService(Executors.newFixedThreadPool(2))
+
+  override def getQueueNames: Array[String] = {
+    Await.result(QueueMetricsOps.getQueueNames(queueManagerActor), timeout.duration).toArray
+  }
+
+  override def getNumberOfMessagesInQueue(queueName: String): CompositeData = {
+    val queueStatistics = Await.result(QueueMetricsOps.getQueueStatistics(queueName, queueManagerActor, new NowProvider), timeout.duration)
+
+    createCompositeDataForQueueStatistics(queueName, queueStatistics)
+  }
+
+  override def getNumberOfMessagesForAllQueues: TabularData = {
+    val queuesStatistics = Await.result(QueueMetricsOps.getQueuesStatistics(queueManagerActor, new NowProvider), timeout.duration)
+    val support = new TabularDataSupport(
+      queueStatisticsTabularType
+    )
+    queuesStatistics.foreach {
+      case (queueName, statistics) => support.put(createCompositeDataForQueueStatistics(queueName, statistics))
+    }
+    support
+  }
+
+  private def createCompositeDataForQueueStatistics(queueName: String, statistics: QueueStatistics) = {
+    val queueDataValues: Array[Object] = Array(
+      queueName,
+      Long.box(statistics.approximateNumberOfVisibleMessages),
+      Long.box(statistics.approximateNumberOfInvisibleMessages),
+      Long.box(statistics.approximateNumberOfMessagesDelayed),
+    )
+
+    new CompositeDataSupport(queueStatisticsCompositeType, queueDataNames, queueDataValues)
+  }
+
+}
+
+object QueuesMetrics {
   val queueDataNames: Array[String] = Array(
     "queueName",
-    "visible",
-    "invisible",
-    "delayed"
+    "ApproximateNumberOfMessages",
+    "ApproximateNumberOfMessagesNotVisible",
+    "ApproximateNumberOfMessagesDelayed"
   )
 
   val queueDataAttributeTypes: Array[OpenType[_]] = Array(
@@ -65,35 +87,4 @@ class QueueMetrics(queueManagerActor: ActorRef) extends QueueMetricsMBean {
     queueStatisticsCompositeType,
     Array("queueName")
   )
-
-  implicit val timeout: Timeout = Timeout(21, TimeUnit.SECONDS)
-  implicit val ec: ExecutionContext = ExecutionContext.fromExecutorService(Executors.newFixedThreadPool(2))
-
-  override def getNumberOfMessagesInQueue(queueName: String): CompositeData = {
-    val queueStatistics = Await.result(QueueMetricsOps.getNumberOfMessagesInQueue(queueName, queueManagerActor, new NowProvider), timeout.duration)
-
-    createCompositeDataForQueueStatistics(queueName, queueStatistics)
-  }
-
-  override def getNumberOfMessagesForAllQueues: TabularData = {
-    val queuesStatistics = Await.result(QueueMetricsOps.getNumberOfMessagesInAllQueues(queueManagerActor, new NowProvider), timeout.duration)
-    val support = new TabularDataSupport(
-      queueStatisticsTabularType
-    )
-    queuesStatistics.foreach {
-      case (queueName, statistics) => support.put(createCompositeDataForQueueStatistics(queueName, statistics))
-    }
-    support
-  }
-
-  private def createCompositeDataForQueueStatistics(queueName: String, statistics: QueueStatistics) = {
-    val queueDataValues: Array[Object] = Array(
-      queueName,
-      Long.box(statistics.approximateNumberOfVisibleMessages),
-      Long.box(statistics.approximateNumberOfInvisibleMessages),
-      Long.box(statistics.approximateNumberOfMessagesDelayed),
-    )
-
-    new CompositeDataSupport(queueStatisticsCompositeType, queueDataNames, queueDataValues)
-  }
 }
